+++ src/extract/rejection.rs
@@ -19,13 +19,6 @@ define_rejection! {
 
 define_rejection! {
     #[status = INTERNAL_SERVER_ERROR]
-    #[body = "URI taken by other extractor"]
-    /// Rejection used if the URI has been taken by another extractor.
-    pub struct UriAlreadyExtracted;
-}
-
-define_rejection! {
-    #[status = INTERNAL_SERVER_ERROR]
     #[body = "Extensions taken by other extractor"]
     /// Rejection used if the method has been taken by another extractor.
     pub struct ExtensionsAlreadyExtracted;
@@ -215,7 +208,6 @@ composite_rejection! {
     /// Contains one variant for each way the [`Query`](super::Query) extractor
     /// can fail.
     pub enum QueryRejection {
-        UriAlreadyExtracted,
         FailedToDeserializeQueryString,
     }
 }
@@ -230,7 +222,6 @@ composite_rejection! {
         FailedToDeserializeQueryString,
         FailedToBufferBody,
         BodyAlreadyExtracted,
-        UriAlreadyExtracted,
         HeadersAlreadyExtracted,
     }
 }